Fix multi thread bug
When paused/resumed quickly, it was possible for the (now outdated)
thread to not be finished, esp. if it was waiting on the long delay
between typing and backspacing. This could cause multiple threads to be
active at once.

Now, each timeout tracks its own ID, and also returns immediately if the
main thread ID has changed.